# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Band & Event Matching API",
    "description": "REST API for matching amateur musicians/bands with events",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "x-api-key",
      "description": "API key for authentication"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/bands/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "city",
            "in": "query",
            "type": "string"
          },
          {
            "name": "state",
            "in": "query",
            "type": "string"
          },
          {
            "name": "tags",
            "in": "query",
            "type": "string"
          },
          {
            "name": "search",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "summary": "Create a band",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Band payload",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Band"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/api/events/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "city",
            "in": "query",
            "type": "string"
          },
          {
            "name": "state",
            "in": "query",
            "type": "string"
          },
          {
            "name": "tags",
            "in": "query",
            "type": "string"
          },
          {
            "name": "search",
            "in": "query",
            "type": "string"
          },
          {
            "name": "status",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "summary": "Create an event",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Event payload",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Event"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Band": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "The Band Name"
        },
        "description": {
          "type": "string",
          "example": "A description of the band"
        },
        "location": {
          "type": "object",
          "properties": {
            "city": {
              "type": "string",
              "example": "Atlanta"
            },
            "state": {
              "type": "string",
              "example": "GA"
            },
            "country": {
              "type": "string",
              "example": "USA"
            },
            "coordinates": {
              "type": "object",
              "properties": {
                "latitude": {
                  "type": "number",
                  "example": 33.749
                },
                "longitude": {
                  "type": "number",
                  "example": -84.388
                }
              }
            }
          }
        },
        "images": {
          "type": "array",
          "example": [
            "https://example.com/image1.jpg",
            "https://example.com/image2.jpg"
          ],
          "items": {
            "type": "string"
          }
        },
        "tags": {
          "type": "array",
          "example": [
            "rock",
            "indie",
            "alternative"
          ],
          "items": {
            "type": "string"
          }
        },
        "previousEvents": {
          "type": "array",
          "example": [],
          "items": {}
        },
        "contactEmail": {
          "type": "string",
          "example": "band@example.com"
        },
        "contactPhone": {
          "type": "string",
          "example": "555-0123"
        },
        "socialLinks": {
          "type": "object",
          "properties": {
            "website": {
              "type": "string",
              "example": "https://example.com"
            },
            "spotify": {
              "type": "string",
              "example": "https://spotify.com/artist/example"
            }
          }
        }
      }
    },
    "Event": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Summer Music Festival"
        },
        "description": {
          "type": "string",
          "example": "An outdoor music festival"
        },
        "location": {
          "type": "object",
          "properties": {
            "venue": {
              "type": "string",
              "example": "Central Park Amphitheater"
            },
            "address": {
              "type": "string",
              "example": "123 Main St"
            },
            "city": {
              "type": "string",
              "example": "Atlanta"
            },
            "state": {
              "type": "string",
              "example": "GA"
            },
            "country": {
              "type": "string",
              "example": "USA"
            },
            "zipCode": {
              "type": "string",
              "example": "30303"
            },
            "coordinates": {
              "type": "object",
              "properties": {
                "latitude": {
                  "type": "number",
                  "example": 33.749
                },
                "longitude": {
                  "type": "number",
                  "example": -84.388
                }
              }
            }
          }
        },
        "date": {
          "type": "string",
          "example": "2025-07-15T00:00:00.000Z"
        },
        "startTime": {
          "type": "string",
          "example": "18:00"
        },
        "endTime": {
          "type": "string",
          "example": "23:00"
        },
        "images": {
          "type": "array",
          "example": [
            "https://example.com/event1.jpg"
          ],
          "items": {
            "type": "string"
          }
        },
        "tags": {
          "type": "array",
          "example": [
            "festival",
            "outdoor",
            "rock"
          ],
          "items": {
            "type": "string"
          }
        },
        "confirmedBands": {
          "type": "array",
          "example": [],
          "items": {}
        },
        "capacity": {
          "type": "number",
          "example": 500
        },
        "ticketPrice": {
          "type": "number",
          "example": 25
        },
        "organizerName": {
          "type": "string",
          "example": "John Doe"
        },
        "organizerEmail": {
          "type": "string",
          "example": "organizer@example.com"
        },
        "status": {
          "type": "string",
          "example": "upcoming"
        }
      }
    }
  },
  "security": [
    {
      "apiKeyAuth": []
    }
  ]
}